perf: improve e2e test performance in CI
- Add npm dependency caching via actions/setup-node cache option
- Add Next.js build cache (packages/web/.next/cache) to avoid full rebuilds
- Parallelize independent build steps (shared-schema + crypto together,
  backend + web together)
- Start backend and web servers simultaneously, wait for both at once
- Increase Playwright workers from 2 to 4 (matches test file count)
- Reduce retries from 2 to 1 to cut time on legitimate failures
- Create Playwright auth setup project with storageState so login runs
  once instead of per-test in authenticated test suites
- Extract authenticated tests into separate *.authenticated.spec.ts files
  that use pre-authenticated browser state
- Reduce job timeout from 30 to 20 minutes
