+++ .gitignore
@@ -8,6 +8,7 @@ node_modules
 
 # testing
 coverage
+**/e2e/.auth/
 
 # next.js
 .next/
